requirements.txt
# Core data engineering libraries
pandas>=2.0.0
pyspark>=3.5.0
pyarrow>=14.0.0

# Cloud and storage
google-cloud-storage>=2.10.0
google-cloud-bigquery>=3.13.0
google-auth>=2.23.0

# Orchestration
prefect>=2.14.0

# Data quality and validation
great-expectations>=0.18.0

# HTTP and API
requests>=2.31.0
aiohttp>=3.9.0

# Data processing
numpy>=1.24.0
python-dateutil>=2.8.0

# Development and utilities
python-dotenv>=1.0.0
tqdm>=4.66.0



